g: implement GPIOs
Add a driver for the GPIO controller.


BRANCH=none
BUG=chrome-os-partner:33816
TEST=press the push buttons on the board and see the console text
message and the LEDs lighting up.

+++ include/gpio.h
@@ -78,7 +78,7 @@ struct gpio_alt_func {
 	uint32_t mask;
 
 	/* Alternate function number */
-	int8_t func;
+	int32_t func;
 
 	/* Module ID (as uint8_t, since enum would be 32-bit) */
 	uint8_t module_id;
